+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.2 (WIP)
 - Add additional vim keybindings: 'g' / 'G' to go to first and last element
+- Add additional keybindings: 'c' to open submission comments in browser
 
 ## 0.1 (2020-11-15)
 - First Release